requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4
    # via derex.runner (setup.py)
attrs==21.2.0
    # via jsonschema
bcrypt==3.2.0
    # via paramiko
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
charset-normalizer==2.0.4
    # via requests
click==8.0.1
    # via
    #   click-plugins
    #   derex.runner (setup.py)
click-plugins==1.1.1
    # via derex.runner (setup.py)
colorama==0.4.4
    # via rich
commonmark==0.9.1
    # via rich
cryptography==3.4.8
    # via paramiko
distro==1.6.0
    # via docker-compose
docker[ssh]==5.0.0
    # via docker-compose
docker-compose==1.29.2
    # via derex.runner (setup.py)
dockerpty==0.4.1
    # via docker-compose
docopt==0.6.2
    # via docker-compose
idna==3.2
    # via requests
importlib-metadata==4.6.4
    # via derex.runner (setup.py)
jinja2==3.0.1
    # via derex.runner (setup.py)
jsonschema==3.2.0
    # via docker-compose
markupsafe==2.0.1
    # via jinja2
paramiko==2.7.2
    # via docker
pluggy==0.13.1
    # via derex.runner (setup.py)
py==1.10.0
    # via derex.runner (setup.py)
pycparser==2.20
    # via cffi
pygments==2.10.0
    # via rich
pymongo==3.12.0
    # via derex.runner (setup.py)
pymysql==1.0.2
    # via derex.runner (setup.py)
pynacl==1.4.0
    # via paramiko
pyrsistent==0.18.0
    # via jsonschema
python-dotenv==0.19.0
    # via docker-compose
pyyaml==5.4.1
    # via
    #   derex.runner (setup.py)
    #   docker-compose
requests==2.26.0
    # via
    #   docker
    #   docker-compose
rich==10.7.0
    # via derex.runner (setup.py)
six==1.16.0
    # via
    #   bcrypt
    #   dockerpty
    #   jsonschema
    #   pynacl
    #   websocket-client
texttable==1.6.4
    # via docker-compose
urllib3==1.26.6
    # via
    #   derex.runner (setup.py)
    #   requests
websocket-client==0.59.0
    # via
    #   docker
    #   docker-compose
zipp==3.5.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
